Very basic repl is working \o/

+++ src/vm/byte_code/chunk.rs
@@ -2,7 +2,7 @@ use super::*;
 use crate::vm::value::Value;
 use std::cmp::Ordering;
 
-pub(crate) type LineNumber = u64;
+pub(crate) type LineNumber = usize;
 pub(crate) type AddressType = usize;
 // start address, end address, line number
 pub(crate) type LineInfo = (AddressType, AddressType, LineNumber);
